pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tapestry"
version = "0.1.0"
description = "Distributed e-ink display controller (Tapestry)"
readme = "README.md"
requires-python = ">=3.13,<4.0"
authors = [
    {name = "Nitz"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Pillow>=9.0.0",
    "requests>=2.25.0",
    "PyYAML>=6.0",
    "flask>=2.0.0",
    "qrcode[pil] (>=8.2,<9.0)",
    "opencv-python>=4.5.0",
    "numpy>=1.20.0",
    "pyzbar (>=0.1.9,<0.2.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.scripts]
tapestry = "tapestry.cli:main"
tapestry-webui = "tapestry.webui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
[tool.poetry.group.dev.dependencies]
qrcode = "^8.2"
opencv-python = "^4.12.0.88"
pytest = "^8.4.2"

